Implement the -DPL_DEPRECATED_f95=ON option
I had previously copied a snapshot of bindings/f95 taken at the start
of the new Fortran binding development to bindings/old_f95, and this
commit similarly copies a snapshot of examples/f95 taken at the start
of the new Fortran binding to examples/old_f95.  In addition, some
modest build-system changes have been made to implement the
-DPL_DEPRECATED_f95=ON option to consistently use this old binding
with the old examples to act as an emergency backup in case the new
Fortran binding does not work for a user and also to extend the life
of the old version of the Fortran binding to be used for old
programmes that are no longer being developed in the slightest.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by running both

scripts/comprehensive_test.sh --cmake_added_options "-DDEFAULT_NO_DEVICES=ON -DPLD_ps=ON  -DDEFAULT_NO_BINDINGS=ON -DENABLE_f95=ON -DPL_DEPRECATED_f95=ON" --do_test_interactive no

and

scripts/comprehensive_test.sh --cmake_added_options "-DDEFAULT_NO_DEVICES=ON -DPLD_ps=ON  -DDEFAULT_NO_BINDINGS=ON -DENABLE_f95=ON" --do_test_interactive no

with the expected results in both cases

-DPL_DEPRECATED_f95=ON:

f95
  Missing examples            :  19
  Differing graphical output  :
  Missing stdout              :
  Differing stdout            :

-DPL_DEPRECATED_f95 not mentioned (which is equivalent to
-DPL_DEPRECATED_f95=OFF, i.e., the new Fortran examples and new
Fortran binding):

f95
  Missing examples            :  19
  Differing graphical output  :  16
  Missing stdout              :
  Differing stdout            :

Example 19 is missing from both comparisons because that example does
not yet work in the latter case and I could not be bothered to
configure different lists of examples for -DPL_DEPRECATED_f95=ON and
-DPL_DEPRECATED_f95=OFF.  For the new bindings and examples Fortran
example 16 has different graphical results than the corresponding C
example because I have implemented the fix for the rectangular issue
that Arjen discovered, but I have stashed that work for this
comparison so that stashed result could be committed and tested
separately after the present commit.