Created a PlDevice class, separate from wxPLDevice, fixed wx aspect problems and text height
Unfortunately this commit ended up including multiple items that I
couldn't disentangle. The initial plan was to create a PlDevice C++
class which can be used as a base class for wxPLDevice and other
future drivers. In particular I plan to split the wxDevice into
two, one where a wxDC is passed in and one which uses the viewer.
However I ended up moving some of the wxWidgets text code around
and then ended up fixing a long standing aspect ratio problem too.

Regarding the PlDevice class, this class is still declared in
wxwidgets.h. It mostly consists of some virtual method declarations
and very little code. The only code I did move into this class was
the code which parses text into lines and line sections. The
conversion to wxString still occurs in this code and it uses
wxCoord, so it is not entirely wxWidgets free, but I will deal with
these later.

The moving around of the text code, made me change how text
transforms are performed and caused me to look into a previous
wxWidgets driver aspect ratio bug. In the end I found that this was
caused by the wxPLplotwindow not being created with the intended
size (it was resized to that size later) and this caused plinit to
be called with the wrong aspect ratio. The wcPLplotwindow class has
had a size parameter added to its constructor to fx this.

I also found that wxWidgets reports the total line spacing as line
height, not just the ascender as assumed in the text size
calculations. This has been fixed.