New Fortran binding: add test of locate mode for example 01
This test only works if the source code is locally modified to set
locate_mode to ".true.".  The call to plGetCursor will have to be changed
to a function reference to help break out of the infinite loop waiting
for events once that subroutine ==> function issue is fixed in the new
Fortran binding.  See FIXME in the updated source code.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the x01f95 target (with locally
modified source code to set locate_mode to .true.) and running both

examples/f95/x01f -dev xwin

and

examples/f95/x01f -dev xcairo

Both gave the expected printed out results you also get for the corresponding C
example (in -locate mode).  However, they both have some issues
with escaping the event loop by one of the two usual methods
which is (1) to hit the escape key or (2) to click on an area
of the plot that is outside the 4 subwindows.

The xwin version was put into an infinite event loop because escape
method (1) is internally transformed to method (2) for this device and
method (2) only works if the above FIXME is fixed.  The xcairo version
implements escape methods (1) and (2) independently.  (2) does not
work for the FIXME reason given, but (1) does work for this device.