New Fortran binding: replace "use" statements within interface blocks by the relevant import statements
Contrary to expectations this change increased build times for the x16f95
target from 3.2 seconds to 3.5 seconds.  But that may be because the
gfortran compiler has been optimised for speed to deal with use
statements but not for import statements, and other compilers
may build more quickly.  In any case I like this style much
better so I will stick with it.