New Fortran binding: replace "includes" by "imports" inside interface blocks
This change produced a slight decrease in compile times, and it
appears to me to be a definite improvement in Fortran style
to get rid of these large numbers of include statements.

As a result of this change, there was no more need for the configured
included_plplot_private_types.f90 file (the file that was included so
many times previously) so I stopped configuring that file for the
build.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the test_diff_psc target in an initially
empty build tree for the shared library case.  There were no build or
run-time errors, and the PostScript diff report for Fortran was as
follows:

f95
  Missing examples            :  19
  Differing graphical output  :
  Missing stdout              :
  Differing stdout            :

That is, there was no change from previous.