New Fortran binding: Make single-precision examples work with double-precision C library
This test was made by temporarily defining plflt to correspond to
kind(1.0) while the C library continued to be double-precision.

One change that was needed was to use plflt real precision as
consistently as possible to avoid the situation where an individual PLplot
routine was called with mixed precision arguments. The typical issue that was
addressed was replacing all dble(...) by real(..., kind=plflt), using
plflt precision variants of the values of the PL_PI, PL_TWOPI, and
PL_NOTSET parameters (which are always double), and replacing all
explicit precision constants (e.g., 1.d0) by the plflt variant.

Another single-precision issue that was addressed was to change the
scaling factor in example 33 from 1.e-200 (which underflows at single
precision) to the maximum of 1.e-200 and an underflow-proof value for
the particular plflt precison being used.  (This is just a workaround,
and eventually that scaling factor should be changed to 1.e-20 for all
language examples to be safe for all real precisions).

A final single-precision issue was that example 29 just cannot work at
single precision because of time representation issues at that
precision.  The solution that I took for this case was to adopt fixed
double precision for example 29, but a more comprehensive approach is
needed in general (e.g., drop example 29 for all languages if the core
library is single precision).

I took this opportunity to update copyright years for Alan W. Irwin
and to use a more modern Fortran idiom in the examples such as
changing all subroutines and functions that are defined in these
examples to CONTAINed procedures and replacing generic unnamed end
statements for programs, subroutines and functions with explicitly
named end statements.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie).  I took advantage of the uniform
nature of the plflt precision of the examples by temporarily adjusting
that to kind(1.) while the C library continued to be double precision.
(So this is the first real test of the new Fortran binding where example
precision is single and C library is double.)  The issues addressed
above were mostly discovered with this special test, but after they
were fixed the examples built and ran without issues and produced
the following difference report compared to the double-precision
C example results:

f95
  Missing examples            :  19
  Differing graphical output  :  04 06 08 09 15 16 18 20 21 22 26 27 28 33
  Missing stdout              :
  Differing stdout            :

I did not look at detailed PostScript diff results, but their
existence is expected since our examples are notorious for being
sensitive to precision issues.  I did eyeball those Fortran single-precision
example/double-precision C library PostScript plots, and they all looked
fine with no obvious visual anomalies.

I then following up by restoring plflt to its expected value of
kind=(1.d0) for this double-precision version of the C library (so both
example and library have the same (double) precision) and
built (starting from an empty build tree) the test_diff_psc
target with no build or run-time issues, and with the following
PostScript difference report:

f95
  Missing examples            :  19
  Differing graphical output  :
  Missing stdout              :
  Differing stdout            :

This is (as expected) the same as when I last ran this standard
version of the test, i.e., the large number of example changes did not
compromise this good result, but as before there is still work to do
on the API associated with example 19.