Add .def file for Intel Fortran on Windows and fix plparseopts
DLLs on Windows require that the symbols that are to be exported are specified
explicitly. This can be done in at least two ways. Here we use the "definition
file" approach.

Using the Intel Fortran and MS Visual C++ compilers revealed a bug in the
wrapper for plparseopts. If no argument was given, it could free an uninitialised
pointer. This has been corrected.

Tested by Arjen Markus (arjenmarkus@users.sourceforge.net) on "bare" Windows and on Cygwin.