New Fortran binding: C routines that return values are interfaced to Fortran functions
Strict application of this rule means that plparseopts and plsetopt
have now both been changed from Fortran subroutines into Fortran
functions with plGetCursor soon to follow.

This is a (now) documented backwards-incompatibility with the new
Fortran binding.  It is documented first because plparseopts is
so ubiquitous and demands users change all use of

call plparseopts(...)

to

integer :: plplarseopts_rc
plparseopts_rc = plparseopts(...)

In addition I dropped the plrandd_single function in the interest of
complying with the KISS principle of always using double precision
entities where there is no automatic way (i.e., via the precision of
arguments supplied by the user) to select the single-precision entity.