New Fortran binding: plmapfill and plmapline implementation
N.B. plmapfill and plmapline are implemented with separate
specifically named functions for the single and double case for the
same reasons as mentioned in the last commits for plmap and
plmeridians.

This commit reduces the Fortran missing API to just plmapstring
(currently untested by any example) and plmaptex (normally tested by example
19)

This addition to the Fortran API now allows running the first 4 pages
of the 19th Fortran example, and the reinstated fifth page as well
(except for the plmaptex calls which have been commented out for that
page for this test).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the x19f95 target and running it with

valgrind examples/f95/x19f -dev svg -fam -o testf.svg

The valgrind report was perfect (0 errors, no leaks are possible), the
first four pages of results were identical to the results from the
corresponding C example, and the fifth page was identical to the
corresponding C result except for the missing text from the commented
out plmaptex calls.  I also temporarily changed plflt to kind(1.) and
ran the same tests with the same good results except for a few
expected minor differences at the +/- 0.03 level between the
single-precision Fortran example results and double-precision example
C results.