New Fortran binding: plmapstring implementation
N.B. plmapstring is implemented with separate specifically named
functions for the single and double case for the same reasons as
mentioned in the last commits for plmapfill, plmapline, plmap, and
plmeridians.

This commit reduces the missing API in the new Fortran binding to
just one function, plmaptex.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the x19f95 target.

N.B. this is only a build test and the actual functionality of
plmapstring at run time was not tested because no example currently
uses plmapstring.