Fortran binding: change plGetCursor into a function
This commit follows two patches sent to me by Arjen whose net effect
is to change plGetCursor from just an interface block describing the C
version of plGetCursor to a full-blown Fortran module procedure that
calls that interfaced C routine. However, this commit goes one step
further and changes that module procedure from a subroutine to a
function and similarly for the interface block since the underlying C
version of plGetCursor is a function that returns a non-zero value on
success and a 0 value on failure.

The use of Fortran module procedure means that Fortran examples 1 and
20 (as confirmed by nm) now no longer have an undefined symbol
plGetCursor that can only be supplied by libplplot.  Instead the
symbol is __plplot_graphics_MOD_plgetcursor_impl which is supplied by
libplplotf95.  As a result, examples 1 and 20 no longer have special
linking needs and can be linked just like the rest of the examples
just to libplplotf95 without it being necessary to explicitly link
libplot. I have made appropriate changes to the CMake-based and
traditional build systems to be consistent with that linking model.

Changing Fortran plGetCursor from a subroutine to a function allows
Fortran example 01 to work correctly when put in locate mode since
that correct working requires the detection of whenever the
return from C plGetCursor and therefore Fortran plGetCursor is equal
to 0 to exit a locate-mode loop.

To Do.  We still need to define private and public versions of the
PLGraphicsIn type argument of plGetCursor.  The private version would
have real(kind=private_plflt) floating-point components and a
NULL-terminated C string component, while the public version would
have real(kind=private_double) (and real(kind=private_single) if there
are no disambiguation issues) floating-point components and a Fortran
character array of the correct length that is right-filled with
blanks.  Then the module procedure would call the interfaced routine
to fill the private version of PLGraphicsIn, then fill a public
version of PLGraphicsIn with the appropriately converted values before
returning that to the calling Fortran procedure.

Tested by: Alan W. Irwin <airwin@users.sourcforge.net> on Linux
(Debian jessie) by running

scripts/comprehensive_test.sh --cmake_added_options "-DDEFAULT_NO_DEVICES=ON -DPLD_ps=ON  -DDEFAULT_NO_BINDINGS=ON -DENABLE_f95=ON" --do_test_interactive no

The summary of differences report is as follows:

f95
  Missing examples            :
  Differing graphical output  :
  Missing stdout              :
  Differing stdout            :

That is, this commit has introduced no regressions in these perfect
test results.

Another test was I ran Fortran example 01 in locate mode (by locally
changing locate_mode from .false. to .true.).  The results were
identical to the corresponding -locate mode C version of the example
for the very first time. That is, clicking with any mouse button or
hitting any key outside a subwindow or hitting the escape key anywhere
on the plot permanently exits the locate-mode loop.

A final test was I ran Fortran example 20 with -dev xwin (the
only device that currently gives interactive results for
this example), and I got the same interactive results as
the C example; the left button allowed me to select an
area of the image with drag and drop and the right button
allowed me to escape that selection loop.