Fortran binding: Final planned improvement for the Fortran version of plGetCursor
The improvement is to use an appropriate public derived type for the argument
of the Fortran plGetCursor function and an appropriate private derived type
for the argument of the interface_plGetCursor function (which is the
Fortran interface to the C version of plGetCursor), and when copying
back results from the private to public derived type do all appropriate
conversions.

The public derived type is called PLGraphicsIn.  It has no bind(c)
attribute, and all the types for its components are the normal public types, i.e.,

* integer,

* string(len=16) :: string (which is a normal blank suffixed Fortran character
string), and

* real(kind=private_double).
(I had to make a choice between private_double and private_single,
and I chose the former for best precision.)

The private derived type is called private_PLGraphicsIn.  It has the
bind(c) attribute and all the types of its components are interoerable
ones, i.e.,

* integer(kind=private_plint),

* character(len=1), dimension(16) :: string, (where string is a
NULL-terminated C string), and

* real(kind=private_plflt)

These characteristics make private_PLGraphicsIn exactly equivalent
to the C struct called PLGraphicsIn which is why this type
of argument is correct for interface_plGetCursor.

I have also updated the DocBook Fortran chapter to reflect this new
definition of the Fortran version of PLGraphicsIn and to document the
constraint that its floating-point arguments are double precision.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian testing) by building a locally modified version of Fortran
example 01 that puts it in locate mode and verifying there was no
regression in the interactive results from that mode with both -dev
xwin and -dev xcairo.  In the latter case, the results were actually
an improvement because of the proper transformation from the
NULL-terminated C string component result for private_PLGraphicsIn to
the Fortran character(len=16) component result for PLGraphicsIn with
correct trailing blanks rather than the trailing garbage we had before.