DocBook documentation: remove ancient ChangeLog file
This Changelog file's last entry was in 2001 and is extraordinarily
incomplete even for the time epoch it does cover.  For those
interested in the historical record of the initial stages of the
implementation of our DocBook-based documentation, the git log
(converted from cvs to svn to git) is a much better reference.