Fix -DPL_DOUBLE=OFF issue: Example 31
Example 31 was erroring out with test failures for at least C and C++
for -DPL_DOUBLE=OFF because floating-point constants (which default to
double for these languages) were being used in argument lists.  I
solved this by replacing constants with PLFLT variables everywhere
possible.  Replacing constants with variables is a much cleaner style
so I propagated that part of the change to all languages regardless
of whether they were strongly typed or not.