Disable example 29 for -DPL_DOUBLE=OFF
Currently this "time" example errors out (even in the C case) with
-DPL_DOUBLE=OFF so temporarily work around the issue by simply not
running this example for noninteractive tests.  Removing the
workaround will be a few-line change if/when a proper fix emerges for
example 29 and -DPL_DOUBLE=OFF. The equivalent temporary workarounds
for interactive tests are possible, but requires substantial changes
to scripts and then substantial changes again to remove the
workarounds.  Therefore we deliberately do not implement the temporary
workarounds in these interactive cases, and instead we will simply
wait for a proper fix to emerge.