Save pls->difilt in plbuf_bop()
The various transformation functions can be called before plinit() which
cause pls->difilt to be set but don't cause pldi_ini() to be called. The
value of pls->difilt is used in plbop() for the auto scaling of text and
tick marks so we need to ensure that it is restored before rdbuf_bop()
calls plbop().