Change epochs used for Example 29 to maximize time-representation precision
Example 29 has been changed for all our supported languages to be much
more careful of adopted epoch to maximize the time-representation
precision.  This is a good illustration of how to use epochs correctly
for our users, and also allows this example to give good results (as
opposed to errors) for the -DPL_DOUBLE=OFF case.

I also solved a time-related build issue for our PLplot library
for the -DPL_DOUBLE=OFF case.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the test_diff_psc target for both the
-DPL_DOUBLE=ON and OFF cases.

For -DPL_DOUBLE=ON the PostScript difference report for the only
languages that were not perfect was as follows:

ada
  Missing examples            :
  Differing graphical output  :  08 19
  Missing stdout              :
  Differing stdout            :
adathick
  Missing examples            :
  Differing graphical output  :  08 19
  Missing stdout              :
  Differing stdout            :
ocaml
  Missing examples            :
  Differing graphical output  :  08 16 19 33
  Missing stdout              :
  Differing stdout            :

i.e., there are no changes here compared to recent previous reports and
therefore no regressions in example 29 results were introduced by these
intrusive changes to that example.

For -DPL_DOUBLE=OFF the PostScript difference report was as follows:

c++
  Missing examples            :
  Differing graphical output  :  20 27 29
  Missing stdout              :
  Differing stdout            :
f95
  Missing examples            :
  Differing graphical output  :  08 16 19 20 27 29
  Missing stdout              :
  Differing stdout            :
tcl
  Missing examples            :
  Differing graphical output  :  08 09 16 19 27 28 29
  Missing stdout              :
  Differing stdout            :

(Note the languages missing from this report are all automatically
disabled for -DPL_DOUBLE=OFF except for Python which I had to manually
disable because of an overflow in one of the examples that I am still
tracking down.) The example 29 differences (according to ndiff -abserr
10) are all small (one unit in the last place) for c++, but there are
a few larger differences in the f95 and tcl results that lead to a
barely visible change to pages 5 and 6 (the 1961-08-01 UTC
discontinuity) for f95 but no visible changes for tcl.  I have
commented on the remaining -DPL_DOUBLE=OFF example differences above
before, and I presume nothing has changed for any of those since
none of those examples have anything to do with the time changes
in this commit.