Solve -DPL_DOUBLE=OFF issues for Python binding and examples
The swig-generated Python binding was originally designed to work
well with both -DPL_DOUBLE=ON and OFF.  This commit restores
good Python results for -DPL_DOUBLE=OFF by fixing the "assumption
that PLFLT was double" bugs that had crept in since this was last looked
at.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the test_diff_psc target.  The
Python and Tcl PostScript difference summary from that result
is as follows:

python
  Missing examples            :
  Differing graphical output  :  08 09 16 19 27 28 29
  Missing stdout              :
  Differing stdout            :
tcl
  Missing examples            :
  Differing graphical output  :  08 09 16 19 27 28 29
  Missing stdout              :
  Differing stdout            :

I include the Tcl results because they are strikingly similar with the
Python results, and in fact deeper investigation shows the Python and
Tcl results are identical with each other (except for a minor rounding
difference for just one example).  These much stronger differences
between either Tcl or Python results and the corresponding C results
are mostly due to lots more rounding errors.  However, for
example 16, the differences with C results go a lot further than that
with whole lines in the PostScript file changed.  However, for example
16 there is nothing visibly different with the gv PostScript viewer.
So I take the above results as a success for Python (and Tcl) for
the -DPL_DOUBLE=OFF case.

But I have two additional conclusions:

(1) I ascribe the near identicality between Python and Tcl results to
both interfacing a double-precision example environment with a
single-precision PLplot library with final PostScript results rounded
the same way in both cases.

(2) I ascribe the much larger differences between those results and
the corresponding C results to the latter being a single-precision
example environment interfacing with a single-precision PLplot
library. Thus, the result of that C example difference is there is
different rounding than in the Python and Tcl example cases.