Clarify copyright language
As discussed some time ago on list, I dropped the language that
essentially stated that everything distributed with PLplot could be
relicensed under the LGPL which is manifestly not true.

Instead, I have emphasized the vast majority of PLplot files are
licensensed for distribution under the LGPL, but there are some
known exceptions to this general rule which are also stated
in the Copyright file.

I reviewed the license notices in the files in bindings/octave and
examples/octave and modified the remarks about Octave file licensing
accordingly.