Update interactive Octave test script
This script runs a large subset of the Octave "p" examples for
interactive devices.  This script is called by "./plplot-test.sh
--interactive_octave" and that command is also called by the
test_octave_xwin target in the build tree and also the CMake-based
installed examples tree.

I updated the script to be exactly consistent with the corresponding
"p" example part of the noninteractive octave test script except for
the list of "p" examples that are run (which is generally much more
complete than for the noninteractive test case).

I only tested the interactive octave test script for interactive
devices xwin and tk because the other interactive devices fail due to
the current well-known stunted interactive capabilities that are
implemented for the other devices compared to xwin and tk.

For the Debian jessie version of octave (3.8.2-4) I discovered a
number of "p" examples that segfaulted (p4, p5, p6, p7, p10, and p11)
so I removed those and p20 (which exercises p4, etc.) from the
complete list of "p" examples.  The result was the interactive octave
tests finally became reliable again with both the xwin and tk devices.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) for the shared case by building the test_octave_xwin in the
build tree and CMake-based installed examples tree and by running

./plplot-test.sh --interactive_octave --device=xwin --verbose

and

./plplot-test.sh --interactive_octave --device=tk --verbose

in the traditionally built installed examples tree. For what it is
worth this was done after running "make test_noninteractive" and "make
test_interactive" from $prefix/share/plplot*/examples, although
building those targets should not be necessary in order to run the
above two test commands.

For all these different tests, there were no run-time errors and the
interactive results (including the new "Chloe" test image) looked good
which verifies, for example, that the single location for Chloe.pgm in
the source tree, build tree, and installed examples tree works for all
these different tests