Minimum CMake version is now 3.0.2 for Linux and 3.3.2 for all other platforms.
The details of the changes are as follows:

The minimum version for Linux stays the same at 3.0.2.

The minimum version for Cygwin has been bumped from 3.0.2 to 3.3.2
(which should be OK since Cygwin started providing CMake-3.3.2 in
2015-11).

The minimum version for all non-Cygwin and non-Linux platforms has
been reduced from 3.4.3 to the previous 3.3.2 at the request
of some of the PLplot developers.

In the future, the rules of thumb I will use to decide when to
bump these versions is as follows:

Linux: the CMake minimum version will be bumped when Debian stable
(currently jessie) provides that version. Debian stable currently
provides 3.0.2.

All other platforms: the CMake minimum version will be bumped when
both Cygwin and MinGW-w64/MSYS2 provide a "matured" version (the last
in any given minor series of CMake releases which is typically
released near the time the next series of minor releases is started).
For CMake 3.x those matured versions are currently 3.0.2, ...,
3.3.2, and 3.4.3. Both Cygwin (see
<http://cygwin.com/cgi-bin2/package-grep.cgi?grep=cmake&arch=x86_64>)
and MinGW-w64/MSYS2 (see
<http://sourceforge.net/projects/msys2/files/REPOS/MINGW/x86_64/>)
currently provide CMake-3.3.2, but not (yet) CMake-3.4.3 which means
our non-Linux minimum of 3.3.2 currently follows this rule of thumb.