Solve a problem with the plplotLibDir variable under "bare" Windows
When using MSVC as the C compiler under Windows, the build for the Tcl binding failed on an unresolved symbol "plplotLibDir". Apparently the combination of macros as used in pldll.h caused this symbol to be marked as "export" instead of "import". With the change in this commit the error disappears and the Tcl examples work fine. This has no consequences for platforms where symbols need not to be decorated with "import" or "export" attributes.

Tested by Arjen Markus <arjenmarkus@users.sourceforge.net> on "bare" Windows (MSVC 2013, 64 bits) and Cygwin (gcc 4.9.3)