First step in implemented redacted API for Tcl
This commit makes it possible to drop most size arguments for the Tcl binding. It does not yet include an adaptation of the examples to this new set-up and testing has been limited so far. Only the generated interfaces have been adapted - for the remaining, more complicated, interfaces not much needs to be done, they already relied on the matrix functions for determining the sizes. The only thing is that some commands are missing one version.

Tested by Arjen Markus <arjenmarkus@users.sourceforge.net> on Cygwin