Revert commit b94370e because it causes build issues
Commit b94370e cannot be correct since it stops essentially all
symbols from being exported from libplplottcltk.  And obviously making
the library useless like this leads to build errors and/or linking
issues for all applications that link to it.  For example, pltcl
requires symbols defined by libplplottcltk to be visible.  I confirmed
build errors for this application on Linux with gcc option
-fvisibility=hidden which are solved by the present commit, but
presumably similar build errors will occur for MSVC prior to
the present commit as well.