Attempt at another fix for the plplotLibDir visibility issue
The fix for this issue that was just reverted essentially dropped
visibility for all symbols in libplplottcltk.  The present fix
instead drops visibility for just the plplotLibDir symbol in
libplplottcltk.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the test_pltcl_standard_examples target.

That custom target built all dependencies correctly which is the
principal positive test result we have at the present time for the
present fix.  The custom target also ran the first 14 examples without
issues, but failed to run the 15th Tcl example because of another
recently introduced issue (the recent plplat changes for the Tcl
binding).  Until those other issues are fixed we necessarily must use
such limited testing of the present fix, but so far so good with the
present fix.