Fix memory management issue for -O2 or -O3 gcc optimization flags
The issue that was fixed was that for the libplplot core library
compiled with gcc optimization flag -O2 or -O3, example 19 produced
valgrind reports indicating memory management issues (invalid reads)
associated with the allocated filename used in src/plmap.c.

So I looked carefully at everything to do with filename and did some
code cleanup (drop attempt to concatanate an empty string, drop
attempt to save truncatedfilename which is never used, and only
calculate warning string when that warning is actually going to be
used).  I also changed the type of filenamelen from int to size_t. I
have no idea which of these minor changes made the difference, but the
result is the valgrind report is now perfect (0 errors, no leaks are
possible) for both gcc options -O2 and -O3.