Automatically generate plplotf95_ifort.def from the symbols in the plplotf95 library
Implement a new (Unix-only) target "check_ifort_definitions_file"
which automatically generates an ifort def file from the symbols in
the latest version of the plplotf95 library and compares that result
with the plplotf95_ifort.def file in the source tree.  If the two
files are not the same, the user has a chance to review the changes.
If those changes are acceptable (the usual case), then the user can
copy by hand the generated file in the build tree on top of the
existing file in the source tree and commit that result.  If the
changes are not acceptable (rare), then tweaks must be made to the
method (a combination of cmake and sed logic) of generating the ifort
def file from the symbols in the plplotf95 library.

For the current case, the changes were to insert a commentary header
(following the ";" comment syntax for .def files documented at
<https://msdn.microsoft.com/en-us/library/h41zhe21.aspx>) and to
adjust to a small number of changed symbol names that are the result
of recent changes (i.e., the plmap* and plmeridians compact style of
implementation) in the way that plplotf95 is implemented.  So although
I cannot test this result (no access to MSVC), my feeling is it should
work so long as the comment syntax that is documented above actually
works for the ifort case.