Document backwards incompatibility in plimagefr signature without callback-related arguments
I implemented this signature for the new Fortran binding some time ago
to be consistent with how the plshade and plshades signatures without
callback-related arguments are handled.  But today I realized this
consistent approach was not compatible with what was implemented (in
error, I believe) in the old Fortran binding. So I have now documented
this backwards incompatibility.