Fortran binding: implement Fortran callbacks for plcont, plshade, plshades, plimagefr, and plvect
This implementation was almost trivial because it used the compact
style, and the required transform callback infrastructure had already
been set up in that style for the plstransform case.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the plplotf95 target.

N.B. This build test was a success, but run-time testing of
all these new Fortran callbacks needs to be done as well.  That
will require changes to the Fortran examples which are still
in the planning stage.