Fortran examples: test all six plcont signatures
Those signatures are (1) no callback-related arguments which means the
pltr0 C "identity transform" callback is used; (2) singly dimensioned
xg and yg which means the pltr1 C callback is used; (3) doubly
dimensioned xg and yg which means the pltr2f C callback is used; (4)
tr which means a special Fortran callback is used that uses the tr
array to do the transformation; (5) Fortran callback without data is
used; and (6) Fortran callback with data is used.

By default the example now tests 2, 3, 5, and 6, but if tr_callback is
locally set to .true., then 5 and 6 are replaced by 4, and if
tr_callback is locally set to .false. and identity_callback locally
set to .true., 5 and 6 are replaced by 1.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the svg, x09c, and x09f95 targets and
running

valgrind examples/c/x09c -dev svg -fam -o testc.svg
valgrind examples/f95/x09f -dev svg -fam -o testf.svg

The valgrind run on x09f was done three times (two with local
modifications of the examples to replace (5 and 6) by either 4 or 1. As a
result, the 6 possible signatures of plcont were tested.

In all cases the valgrind reports were perfect (0 errors, no leaks
possible).  Furthermore, the Fortran and C svg results were identical
when replacing (5 and 6) by 4.  As expected the svg result has
different tick marks and labels of those tick marks, but not different
contours when replacing (5 and 6) by 1 since the latter (identity
transform) case maps to the range of x and y indices, rather than the
intervals of -1. to 1. that are obtained with either (5 and 6) or 4.

In sum, these tests of the six possible plcont signatures showed the
expected results in all cases.