Fix long-standing off-by-one errors in plimage and plimagefr
Prior to this fix the selected subset of Chloe image (page 3 of
example 20) was displayed with a small offset (as determined by blink
comparing the two) from the original Chloe image (page 2 of the
example).  This fix cures this offset and also maintains the good
alignment between the original image and the reduced dynamic range
variant of this image (page 5 of the example).