Fortran binding: implement "tr" signature of plimagefr
This commit rectifies an oversight in the implementation of the new
Fortran binding which should for backwards compatibility reasons
supply a "tr" signature for plimagefr since that signature was also
supplied by the old Fortran binding.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the plplotf95 library without issues.

N.B. this is just a build test, and the implementation of a run-time
test for this plimagefr signature is planned soon.