Fix PostScript difference regressions introduced by dde39c8c (Fix long-standing off-by-one errors in plimage and plimagefr)
Commit dde39c8c was not sufficiently tested so that PostScript
difference regressions were only discovered after the commit was made.
These regressions (discovered by building the test_diff_psc target)
were that page 5 of example 20 was not rendered consistently with the
C result for the Python, Tcl, and Java languages.

These regressions were caused by pltr0 no longer having the identical
effect as a NULL callback as a result of the off-by-one fixes.  The
solutions for these regressions are to use a NULL argument rather than
pltr0 for the C callback when the example has no callback-related
arguments in the call to plimagefr (this change was made to the Python
and Tcl bindings), and since the Java bindings do not currently have
an option to generate a NULL argument for the C callback modify the xg
and yg arrays in the Java example to exactly mimic the effect of a
NULL argument for the C callback.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the test_diff_psc target.  As a result
of the above fixes the example 20 PostScript difference regressions are
now gone.  The remaining differences (which have been with us for some
time) are the following:

tcl
  Missing examples            :
  Differing graphical output  :  15 18
  Missing stdout              :
  Differing stdout            :
ada
  Missing examples            :
  Differing graphical output  :  08 19
  Missing stdout              :
  Differing stdout            :
adathick
  Missing examples            :
  Differing graphical output  :  08 19
  Missing stdout              :
  Differing stdout            :
ocaml
  Missing examples            :
  Differing graphical output  :  08 16 19 33
  Missing stdout              :
  Differing stdout            :